// database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json }
  | Json[]

export interface Database {
// ...
  public: {
    Tables: {
      categories: {
        Row: {
          created_at: string
          id: string
          title: string
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          title: string
          user_id: string
        }
        Update: {
          created_at?: string
          id?: string
          title?: string
          user_id?: string
        }
      }
      ingredients: {
        Row: {
          category: string
          created_at: string
          id: string
          lot_size: string
          lot_unit: string
          price: number
          quantity: number
          salt: number | null
          supplier: string | null
          title: string
          unit: string
          user_id: string
        }
        Insert: {
          category: string
          created_at?: string
          id?: string
          lot_size: string
          lot_unit: string
          price: number
          quantity: number
          salt?: number | null
          supplier?: string | null
          title: string
          unit: string
          user_id: string
        }
        Update: {
          category?: string
          created_at?: string
          id?: string
          lot_size?: string
          lot_unit?: string
          price?: number
          quantity?: number
          salt?: number | null
          supplier?: string | null
          title?: string
          unit?: string
          user_id?: string
        }
      }
      recipes: {
        Row: {
          created_at: string
          id: string
          image: string | null
          ingredients: Json
          instruction: string | null
          price: number
          title: string
        }
        Insert: {
          created_at?: string
          id?: string
          image?: string | null
          ingredients: Json
          instruction?: string | null
          price: number
          title: string
        }
        Update: {
          created_at?: string
          id?: string
          image?: string | null
          ingredients?: Json
          instruction?: string | null
          price?: number
          title?: string
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}